+++ csvdiff/__init__.py
@@ -12,126 +12,72 @@ __email__ = 'lars@yencken.org'
 __version__ = '0.2.0'
 
 import sys
-import csv
-import json
 
 import click
 
+from . import records, patch, error
 
-DEBUG = False
-
-
-class FatalError(Exception):
-    pass
-
-
-def csvdiff(lhs, rhs, index_columns):
-    lhs_recs = load_records(lhs, index_columns)
-    rhs_recs = load_records(rhs, index_columns)
-
-    orig_size = len(lhs_recs)
-
-    return diff_records(lhs_recs, rhs_recs), orig_size
+# exit codes for the command-line
+EXIT_SAME = 0
+EXIT_DIFFERENT = 1
+EXIT_ERROR = 2
 
 
-def diff_records(lhs_recs, rhs_recs):
-    # examine keys for overlap
-    removed, added, shared = diff_keys(lhs_recs, rhs_recs)
-
-    # check for changed rows
-    changed = diff_shared(lhs_recs, rhs_recs, shared)
-
-    # summarize changes
-    diff = diff_summary(removed, added, changed, lhs_recs, rhs_recs)
-
-    return diff
-
-
-def load_records(filename, index_columns):
-    try:
-        return {
-            tuple(r[i] for i in index_columns): r
-            for r in csv.DictReader(open(filename))
-        }
-    except KeyError as k:
-        abort('invalid column name {k} as key'.format(k=k))
+def diff_files(from_file, to_file, index_columns):
+    """
+    Diff two CSV files, returning the patch which transforms one into the
+    other.
+    """
+    with open(from_file) as from_stream:
+        with open(to_file) as to_stream:
+            from_records = records.load(from_stream)
+            to_records = records.load(to_stream)
+            return patch.create(from_records, to_records, index_columns)
 
 
-def diff_keys(lhs_recs, rhs_recs):
-    lhs_keys = set(lhs_recs)
-    rhs_keys = set(rhs_recs)
-    removed = lhs_keys.difference(rhs_keys)
-    shared = lhs_keys.intersection(rhs_keys)
-    added = rhs_keys.difference(lhs_keys)
-    return removed, added, shared
+def diff_records(from_records, to_records, index_columns):
+    """
+    Diff two sequences of dictionary records, returning the patch which
+    transforms one into the other.
+    """
+    return patch.create(from_records, to_records, index_columns)
 
 
-def diff_shared(lhs_recs, rhs_recs, keys):
-    "Return the set of keys which have changed."
-    return set(
-        k for k in keys
-        if sorted(lhs_recs[k].items()) != sorted(rhs_recs[k].items())
-    )
+def patch_file(patch_stream, fromcsv_stream, tocsv_stream, strict=True):
+    """
+    Apply the patch to the source CSV file, and save the result to the target
+    file.
+    """
+    diff = patch.load(patch_stream)
 
+    from_records = records.load(fromcsv_stream)
+    to_records = patch.apply(diff, from_records, strict=strict)
 
-def diff_summary(removed, added, changed, lhs_recs, rhs_recs):
-    diff = {}
-    diff[u'removed'] = [lhs_recs[k] for k in removed]
-    diff[u'added'] = [rhs_recs[k] for k in added]
-    diff[u'changed'] = [{'key': k,
-                         'fields': rec_diff(lhs_recs[k], rhs_recs[k])}
-                        for k in changed]
-    return diff
-
-
-def rec_diff(lhs, rhs):
-    "Diff an individual row."
-    delta = {}
-    for k in set(lhs).union(rhs):
-        from_ = lhs[k]
-        to_ = rhs[k]
-        if from_ != to_:
-            delta[k] = {'from': from_, 'to': to_}
-
-    return delta
-
-
-def summarize_diff(diff, orig_size, stream=sys.stdout):
-    if orig_size == 0:
-        # slightly arbitrary when the original data was empty
-        orig_size = 1
-
-    n_removed = len(diff['removed'])
-    n_added = len(diff['added'])
-    n_changed = len(diff['changed'])
-
-    if n_removed or n_added or n_changed:
-        print(u'%d rows removed (%.01f%%)' % (
-            n_removed, 100 * n_removed / orig_size
-        ), file=stream)
-        print(u'%d rows added (%.01f%%)' % (
-            n_added, 100 * n_added / orig_size
-        ), file=stream)
-        print(u'%d rows changed (%.01f%%)' % (
-            n_changed, 100 * n_changed / orig_size
-        ), file=stream)
+    # what order should the columns be in?
+    if to_records:
+        # have data, use a nice ordering
+        all_columns = to_records[0].keys()
+        index_columns = diff['_index']
+        fieldnames = _nice_fieldnames(all_columns, index_columns)
     else:
-        print('files are identical')
+        # no data, use the original order
+        fieldnames = from_records.fieldnames
+
+    records.save(to_records, fieldnames, tocsv_stream)
 
 
-def json_diff(diff, stream=sys.stdout, compact=False):
-    if compact:
-        json.dump(diff, stream)
-    else:
-        json.dump(diff, stream, indent=2, sort_keys=True)
+def patch_records(diff, from_records, strict=True):
+    """
+    Apply the patch to the sequence of records, returning the transformed
+    records.
+    """
+    return patch.apply(diff, from_records, strict=strict)
 
 
-def abort(message=None):
-    if DEBUG:
-        raise FatalError(message)
-
-    print(message, file=sys.stderr)
-    sys.exit(1)
+def _nice_fieldnames(all_columns, index_columns):
+    "Indexes on the left, other fields in alphabetical order on the right."
+    non_index_columns = set(all_columns).difference(index_columns)
+    return index_columns + sorted(non_index_columns)
 
 
 class CSVType(click.ParamType):
@@ -167,27 +113,106 @@ class CSVType(click.ParamType):
                     'or give a summary instead'))
 @click.option('--output', '-o', type=click.Path(),
               help='Output to a file instead of stdout')
-def main(index_columns, from_csv, to_csv, style=None, output=None):
+def csvdiff_cmd(index_columns, from_csv, to_csv, style=None, output=None):
     """
     Compare two csv files to see what rows differ between them. The files
     are each expected to have a header row, and for each row to be uniquely
     identified by one or more indexing columns.
     """
-    csvdiff_main(index_columns, from_csv, to_csv, style=style, output=output)
+    ostream = (sys.stdout
+               if output is None
+               else open(output, 'w'))
+
+    try:
+        if style == 'summary':
+            _diff_and_summarize(from_csv, to_csv, index_columns, ostream)
+        else:
+            compact = (style == 'compact')
+            _diff_files_to_stream(from_csv, to_csv, index_columns, ostream,
+                                  compact=compact)
+
+    except records.InvalidKeyError as e:
+        error.abort(e.args[0])
+
+    finally:
+        ostream.close()
 
 
-def csvdiff_main(index_columns, from_csv, to_csv, style=None, output=None):
-    diff, orig_size = csvdiff(from_csv, to_csv, index_columns)
+def _diff_files_to_stream(from_csv, to_csv, index_columns, ostream,
+                          compact=False):
+    diff = diff_files(from_csv, to_csv, index_columns)
+    patch.save(diff, ostream, compact=compact)
+    exit_code = (EXIT_SAME
+                 if patch.is_empty(diff)
+                 else EXIT_DIFFERENT)
+    sys.exit(exit_code)
 
-    if output is None:
-        ostream = sys.stdout
+
+def _diff_and_summarize(from_csv, to_csv, index_columns, stream=sys.stdout):
+    """
+    Print a summary of the difference between the two files.
+    """
+    from_records = records.load(from_csv)
+    to_records = records.load(to_csv)
+    diff = patch.create(from_records, to_records, index_columns)
+    _summarize_diff(diff, len(from_records), stream=stream)
+    exit_code = (EXIT_SAME
+                 if patch.is_empty(diff)
+                 else EXIT_DIFFERENT)
+    sys.exit(exit_code)
+
+
+def _summarize_diff(diff, orig_size, stream=sys.stdout):
+    if orig_size == 0:
+        # slightly arbitrary when the original data was empty
+        orig_size = 1
+
+    n_removed = len(diff['removed'])
+    n_added = len(diff['added'])
+    n_changed = len(diff['changed'])
+
+    if n_removed or n_added or n_changed:
+        print(u'%d rows removed (%.01f%%)' % (
+            n_removed, 100 * n_removed / orig_size
+        ), file=stream)
+        print(u'%d rows added (%.01f%%)' % (
+            n_added, 100 * n_added / orig_size
+        ), file=stream)
+        print(u'%d rows changed (%.01f%%)' % (
+            n_changed, 100 * n_changed / orig_size
+        ), file=stream)
     else:
-        ostream = open(output, 'w')
+        print('files are identical')
 
-    if style == 'summary':
-        summarize_diff(diff, orig_size, ostream)
-    else:
-        compact = (style == 'compact')
-        json_diff(diff, ostream, compact=compact)
 
-    ostream.close()
+@click.command()
+@click.argument('input_csv', type=click.Path(exists=True))
+@click.option('--input', '-i', type=click.Path(exists=True),
+              help='Read the JSON patch from the given file.')
+@click.option('--output', '-o', type=click.Path(),
+              help='Write the transformed CSV to the given file.')
+@click.option('--strict/--no-strict', default=True,
+              help='Whether or not to tolerate a changed source document '
+                   '(default: strict)')
+def csvpatch_cmd(input_csv, input=None, output=None, strict=True):
+    """
+    Apply the changes from a csvdiff patch to an existing CSV file.
+    """
+    patch_stream = (sys.stdin
+                    if input is None
+                    else open(input))
+    tocsv_stream = (sys.stdout
+                    if output is None
+                    else open(output, 'w'))
+    fromcsv_stream = open(input_csv)
+
+    try:
+        patch_file(patch_stream, fromcsv_stream, tocsv_stream, strict=strict)
+
+    except patch.InvalidPatchError as e:
+        error.abort('reading patch, {0}'.format(e.args[0]))
+
+    finally:
+        patch_stream.close()
+        fromcsv_stream.close()
+        tocsv_stream.close()